Fix column style menus for string & date(time)
