+++ rust/perspective-viewer/src/rust/components/style/style_cache.rs
@@ -56,9 +56,7 @@ impl StyleCache {
             let first = map.values().next().cloned();
             map.insert(name, style.clone());
             if let Some(x) = first {
-                if x.is_connected() {
-                    x.get_root_node().insert_before(&style, Some(&x)).unwrap();
-                }
+                x.get_root_node().insert_before(&style, Some(&x)).unwrap();
             }
         }
     }
